pyproject.toml
[project]
name = "python-config-template"
version = "0.1.0"

requires-python = ">=3.13"
dependencies = ["pydantic>=2.11.4"]

[dependency-groups]
shared = ["azure-storage-blob>=12.21.0,<13.0.0", "tqdm>=4.67.1,<5.0.0"]

test = [
    "pytest>=8.3.5",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=6.1.1",
    "pytest-env>=1.1.5",
    "pytest-only>=2.1.2",
    "pytest-picked>=0.5.1",
]

dev = [
    "mypy>=1.9.0,<2.0.0",
    "nbstripout>=0.7.1,<1.0.0",
    "pre-commit>=3.5.0,<4.0.0",
    "python-dotenv[cli]>=1.0.1,<2.0.0",
    "ruff>=0.5.4,<1.0.0",
]

typing = ["types-requests>=2.32.0.20240712,<3.0.0"]

notebook = ["ipykernel>=6.29.4,<7.0.0"]

[tool.uv]
package = false
default-groups = "all"
